+++ src/sqlfluff/dialects/dialect_ansi.py
@@ -26,7 +26,7 @@ ansi_dialect = Dialect('ansi')
 ansi_dialect.set_lexer_struct([
     # name, type, pattern, kwargs
     ("whitespace", "regex", r"[\t ]+", dict(type='whitespace')),
-    ("inline_comment", "regex", r"(--|#)[^\n]*", dict(is_comment=True, type='comment')),
+    ("inline_comment", "regex", r"(--|#)[^\n]*", dict(is_comment=True, type='comment', trim_start=('--', '#'))),
     (
         "block_comment", "regex", r"\/\*([^\*]|\*(?!\/))*\*\/",
         dict(
@@ -112,7 +112,7 @@ ansi_dialect.add(
     NakedIdentifierSegment=SegmentGenerator(
         # Generate the anti template from the set of reserved keywords
         lambda dialect: ReSegment.make(
-            r"[A-Z0-9_]*[A-Z][A-Z0-9_]*", name='identifier', type='naked_identifier',
+            r"[A-Z0-9_]*[A-Z][A-Z0-9_]*", name='naked_identifier', type='identifier',
             _anti_template=r"^(" + r'|'.join(dialect.sets('reserved_keywords')) + r")$")
     ),
     FunctionNameSegment=ReSegment.make(r"[A-Z][A-Z0-9_]*", name='function_name', type='function_name'),
@@ -124,11 +124,11 @@ ansi_dialect.add(
             r"^(" + r"|".join(dialect.sets('datetime_units')) + r")$",
             name='date_part', type='date_part')
     ),
-    QuotedIdentifierSegment=NamedSegment.make('double_quote', name='identifier', type='quoted_identifier'),
-    QuotedLiteralSegment=NamedSegment.make('single_quote', name='literal', type='quoted_literal'),
-    NumericLiteralSegment=NamedSegment.make('numeric_literal', name='literal', type='numeric_literal'),
-    TrueSegment=KeywordSegment.make('true', name='true', type='boolean_literal'),
-    FalseSegment=KeywordSegment.make('false', name='false', type='boolean_literal'),
+    QuotedIdentifierSegment=NamedSegment.make('double_quote', name='quoted_identifier', type='identifier'),
+    QuotedLiteralSegment=NamedSegment.make('single_quote', name='quoted_literal', type='literal'),
+    NumericLiteralSegment=NamedSegment.make('numeric_literal', name='numeric_literal', type='literal'),
+    TrueSegment=KeywordSegment.make('true', name='boolean_literal', type='literal'),
+    FalseSegment=KeywordSegment.make('false', name='boolean_literal', type='literal'),
     # We use a GRAMMAR here not a Segment. Otherwise we get an unecessary layer
     SingleIdentifierGrammar=OneOf(Ref('NakedIdentifierSegment'), Ref('QuotedIdentifierSegment')),
     BooleanLiteralGrammar=OneOf(Ref('TrueSegment'), Ref('FalseSegment')),
@@ -231,6 +231,23 @@ class ObjectReferenceSegment(BaseSegment):
         code_only=False
     )
 
+    def iter_raw_references(self):
+        """Generate a list of reference strings and elements.
+
+        Each element is a tuple of (str, segment). If some are
+        split, then a segment may appear twice, but the substring
+        will only appear once.
+        """
+        # Extract the references from those identifiers (because some may be quoted)
+        for elem in self.recursive_crawl('identifier'):
+            # trim on quotes and split out any dots.
+            for part in elem.raw_trimmed().split('.'):
+                yield part, elem
+
+    def is_qualified(self):
+        """Return if there is more than one element to the reference."""
+        return len(list(self.iter_raw_references())) > 1
+
 
 @ansi_dialect.segment()
 class ArrayAccessorSegment(BaseSegment):
@@ -243,13 +260,9 @@ class ArrayAccessorSegment(BaseSegment):
     )
     parse_grammar = Bracketed(
         Delimited(
-            # We use the no-match version here so we get the correct
-            # handling of the potential colon. We also use Delimited
-            # so that we look for the colon FIRST, because we should
-            # know to expect one and the parser gets confused otherwise.
             OneOf(
                 Ref('NumericLiteralSegment'),
-                Ref('ExpressionSegment_NoMatch')
+                Ref('ExpressionSegment')
             ),
             delimiter=Ref('SliceSegment')
         ),
@@ -305,7 +318,7 @@ class QualifiedNumericLiteralSegment(BaseSegment):
 
 ansi_dialect.add(
     # FunctionContentsExpressionGrammar intended as a hook to override
-    # in other dialects
+    # in other dialects.
     FunctionContentsExpressionGrammar=Ref('ExpressionSegment'),
     FunctionContentsGrammar=OneOf(
         # A Cast-like function
@@ -334,8 +347,35 @@ ansi_dialect.add(
                 ),
             ),
         )
-    )
+    ),
+    # Optional OVER suffix for window functions.
+    # This is supported in biquery & postgres (and it's derivatives)
+    # and so is included here for now.
+    PostFunctionGrammar=Ref('OverClauseSegment')
 )
+
+
+@ansi_dialect.segment()
+class OverClauseSegment(BaseSegment):
+    """An OVER clause for window functions."""
+    type = 'over_clause'
+    match_grammar = Sequence(
+        'OVER',
+        Bracketed(
+            Anything(optional=True)
+        ),
+    )
+
+    parse_grammar = Sequence(
+        'OVER',
+        Bracketed(
+            Sequence(
+                Ref('PartitionClauseSegment', optional=True),
+                Ref('OrderByClauseSegment', optional=True),
+                Ref('FrameClauseSegment', optional=True)
+            )
+        ),
+    )
 
 
 @ansi_dialect.segment()
@@ -355,14 +395,9 @@ class FunctionSegment(BaseSegment):
                 Anything(optional=True)
             ),
         ),
-        Sequence(
-            'OVER',
-            Bracketed(
-                Anything(optional=True)
-            ),
-            optional=True
-        )
+        Ref('PostFunctionGrammar', optional=True)
     )
+
     parse_grammar = Sequence(
         Sequence(
             Ref('FunctionNameSegment'),
@@ -373,19 +408,7 @@ class FunctionSegment(BaseSegment):
                     optional=True)
             ),
         ),
-        # Optional suffix for window functions.
-        # TODO: Should this be in a different dialect?
-        Sequence(
-            'OVER',
-            Bracketed(
-                Sequence(
-                    Ref('PartitionClauseSegment', optional=True),
-                    Ref('OrderByClauseSegment', optional=True),
-                    Ref('FrameClauseSegment', optional=True)
-                )
-            ),
-            optional=True
-        )
+        Ref('PostFunctionGrammar', optional=True)
     )
 
 
@@ -427,6 +450,17 @@ class FrameClauseSegment(BaseSegment):
     # )
 
 
+ansi_dialect.add(
+    # This is a hook point to allow subclassing for other dialects
+    PostTableExpressionGrammar=Sequence(
+        'WITH',
+        'OFFSET',
+        'AS',
+        Ref('SingleIdentifierGrammar'),
+    ),
+)
+
+
 @ansi_dialect.segment()
 class TableExpressionSegment(BaseSegment):
     """A table expression."""
@@ -450,14 +484,33 @@ class TableExpressionSegment(BaseSegment):
             # Values clause?
         ),
         Ref('AliasExpressionSegment', optional=True),
-        Sequence(
-            'WITH',
-            'OFFSET',
-            'AS',
-            Ref('SingleIdentifierGrammar'),
-            optional=True
-        ),
+        Ref('PostTableExpressionGrammar', optional=True)
     )
+
+    def get_eventual_alias(self):
+        """Return the eventual table name referred to by this table expression.
+
+        Returns:
+            :obj:`tuple` of (:obj:`str`, :obj:`BaseSegment`, :obj:`bool`) containing
+                a string representation of the alias, a reference to the
+                segment containing it, and whether it's an alias.
+
+        """
+        alias_expression = self.get_child('alias_expression')
+        if alias_expression:
+            # If it has an alias, return that
+            segment = alias_expression.get_child('identifier')
+            return (segment.raw, segment, True)
+
+        # If not return the object name (or None if there isn't one)
+        ref = self.get_child('object_reference')
+        if ref:
+            # Return the last element of the reference, which
+            # will already be a tuple.
+            penultimate_ref = list(ref.iter_raw_references())[-1]
+            return (*penultimate_ref, False)
+        # No references or alias, return None
+        return None
 
 
 @ansi_dialect.segment()
@@ -489,11 +542,7 @@ class SelectTargetElementSegment(BaseSegment):
         ),
         Sequence(
             OneOf(
-                # We use the unbound version here, so that we can optionally
-                # have space for our Alias at the end. This is potentially
-                # very slow. There's probably a better way for this, but
-                # it's not obvious what that is right now.
-                Ref('ExpressionSegment_NoMatch'),
+                Ref('ExpressionSegment'),
             ),
             Ref('AliasExpressionSegment', optional=True)
         ),
@@ -525,6 +574,23 @@ class SelectClauseSegment(BaseSegment):
         ),
         Dedent
     )
+
+
+# We define the grammar seperately here because it's used in both the
+# parsing and matching routines of the JoinClauseSegment.
+InitialJoinGrammar = Sequence(
+    # NB These qualifiers are optional
+    AnyNumberOf(
+        Ref('FullKeywordSegment'),
+        Ref('InnerKeywordSegment'),
+        Ref('LeftKeywordSegment'),
+        Ref('CrossKeywordSegment'),
+        max_times=1,
+        optional=True
+    ),
+    Ref('OuterKeywordSegment', optional=True),
+    Ref('JoinKeywordSegment')
+)
 
 
 @ansi_dialect.segment()
@@ -576,6 +642,17 @@ class JoinClauseSegment(BaseSegment):
         Dedent
     )
 
+    def get_eventual_alias(self):
+        """Return the eventual table name referred to by this join clause."""
+        table_expression = self.get_child('table_expression')
+        return table_expression.get_eventual_alias()
+
+
+ansi_dialect.add(
+    # This is a hook point to allow subclassing for other dialects
+    JoinLikeClauseGrammar=Nothing()
+)
+
 
 @ansi_dialect.segment()
 class FromClauseSegment(BaseSegment):
@@ -598,13 +675,7 @@ class FromClauseSegment(BaseSegment):
             # Optional old school delimited joins
             Ref('TableExpressionSegment'),
             delimiter=Ref('CommaSegment'),
-            terminator=OneOf(
-                'JOIN',
-                'CROSS',
-                'INNER',
-                'LEFT',
-                'FULL'
-            )
+            terminator=Ref('JoinClauseSegment')
         ),
         # NB: The JOIN clause is *part of* the FROM clause
         # and so should be on a sub-indent of it. That isn't
@@ -615,44 +686,51 @@ class FromClauseSegment(BaseSegment):
         Dedent.when(indented_joins=False),
         AnyNumberOf(
             Ref('JoinClauseSegment'),
+            Ref('JoinLikeClauseGrammar'),
             optional=True
         ),
         Dedent.when(indented_joins=True)
     )
+
+    def get_eventual_aliases(self):
+        """List the eventual aliases of this from clause.
+
+        Comes as a list of tuples (string, segment).
+        """
+        buff = []
+        direct_table_children = self.get_children('table_expression')
+        join_clauses = self.get_children('join_clause')
+        # Iterate through the potential sources of aliases
+        for clause in (*direct_table_children, *join_clauses):
+            ref = clause.get_eventual_alias()
+            # Only append if non null. A None reference, may
+            # indicate a generator expression or similar.
+            if ref:
+                buff.append(ref)
+        return buff
 
 
 @ansi_dialect.segment()
 class CaseExpressionSegment(BaseSegment):
     """A `CASE WHEN` clause."""
     type = 'case_expression'
-    # This method of matching doesn't work with nested case statements.
-    # TODO: Develop something more powerful for this.
-    # match_grammar = StartsWith(
-    #     'CASE',
-    #     terminator='ED'),
-    #     include_terminator=True
-    # )
     match_grammar = Sequence(
         'CASE',
         Indent,
         AnyNumberOf(
             Sequence(
-                # We use the unbound version of Expression here, so that we
-                # deal with potentially nested case statements where the
-                # parsing gets confused by which WHERE and END goes with
-                # which CASE. TODO: Come up with a better solution for this.
                 'WHEN',
                 Indent,
-                Ref('ExpressionSegment_NoMatch'),
+                Ref('ExpressionSegment'),
                 'THEN',
-                Ref('ExpressionSegment_NoMatch'),
+                Ref('ExpressionSegment'),
                 Dedent
             )
         ),
         Sequence(
             'ELSE',
             Indent,
-            Ref('ExpressionSegment_NoMatch'),
+            Ref('ExpressionSegment'),
             Dedent,
             optional=True
         ),
@@ -771,64 +849,19 @@ class ExpressionSegment(BaseSegment):
     """A expression, either arithmetic or boolean.
 
     NB: This is potentially VERY recursive and
-    mostly uses the grammars above.
-    """
-    type = 'expression'
-    match_grammar = GreedyUntil(
-        Ref('CommaSegment'),
-        'AS',
-        'ASC',
-        'DESC',
-        'INNER',
-        'LEFT',
-        'CROSS',
-        'JOIN',
-        'WHERE',
-        'GROUP',
-        'ORDER',
-    )
-    parse_grammar = Ref('Expression_A_Grammar')
 
-
-@ansi_dialect.segment()
-class ExpressionSegment_NoMatch(ExpressionSegment):
-    """A expression, either arithmetic or boolean.
-
-    NB: This is potentially VERY recursive and
     mostly uses the grammars above. This version
     also doesn't bound itself first, and so is potentially
     VERY SLOW. I don't really like this solution.
 
-    The purpose of this particular version of the segment
-    is so that we can make sure we don't swallow a potential
-    alias following the epxression. The other expression
-    segments are more efficient but potentially parse
-    alias expressions incorrectly if no AS keyword is used.
+    We rely on elements of the expression to bound
+    themselves rather than bounding at the expression
+    level. Trying to bound the ExpressionSegment itself
+    has been too unstable and not resilient enough to
+    other bugs.
     """
+    type = 'expression'
     match_grammar = Ref('Expression_A_Grammar')
-    parse_grammar = None
-
-
-@ansi_dialect.segment()
-class ExpressionSegment_TermWhenElse(ExpressionSegment):
-    """Expression terminated by WHEN, END or ELSE."""
-    match_grammar = GreedyUntil(
-        'WHEN',
-        'ELSE',
-        'END'
-    )
-
-
-@ansi_dialect.segment()
-class ExpressionSegment_TermThen(ExpressionSegment):
-    """Expression terminated by THEN."""
-    match_grammar = GreedyUntil('THEN')
-
-
-@ansi_dialect.segment()
-class ExpressionSegment_TermEnd(ExpressionSegment):
-    """Expression terminated by END."""
-    match_grammar = GreedyUntil('END')
 
 
 @ansi_dialect.segment()