+++ src/sqlfluff/dialects/dialect_bigquery.py
@@ -23,7 +23,7 @@ bigquery_dialect.patch_lexer_struct([
 ])
 
 bigquery_dialect.add(
-    DoubleQuotedLiteralSegment=NamedSegment.make('double_quote', name='literal', type='quoted_literal')
+    DoubleQuotedLiteralSegment=NamedSegment.make('double_quote', name='quoted_literal', type='literal', trim_chars=('"',))
 )
 
 # Add the microsecond unit
@@ -50,7 +50,7 @@ class IntervalExpressionSegment(BaseSegment):
 
 
 bigquery_dialect.replace(
-    QuotedIdentifierSegment=NamedSegment.make('back_quote', name='identifier', type='quoted_identifier'),
+    QuotedIdentifierSegment=NamedSegment.make('back_quote', name='quoted_identifier', type='identifier', trim_chars=('`',)),
     IntervalExpressionSegment=IntervalExpressionSegment,
     LiteralGrammar=OneOf(
         Ref('QuotedLiteralSegment'), Ref('DoubleQuotedLiteralSegment'), Ref('NumericLiteralSegment'),