+++ src/sqlfluff/linter.py
@@ -7,6 +7,7 @@ import logging
 
 from difflib import SequenceMatcher
 from benchit import BenchIt
+import pathspec
 
 from .errors import SQLLexError, SQLParseError, SQLTemplaterError
 from .parser import FileSegment, ParseContext
@@ -19,7 +20,7 @@ from .cli.formatters import (format_linting_path, format_file_violations,
                              format_filename, format_config_vals)
 
 
-class LintedFile(namedtuple('ProtoFile', ['path', 'violations', 'time_dict', 'tree', 'file_mask'])):
+class LintedFile(namedtuple('ProtoFile', ['path', 'violations', 'time_dict', 'tree', 'file_mask', 'ignore_mask'])):
     """A class to store the idea of a linted file."""
     __slots__ = ()
 
@@ -31,7 +32,7 @@ class LintedFile(namedtuple('ProtoFile', ['path', 'violations', 'time_dict', 'tr
         If they don't then this function raises that error.
         """
         vs = []
-        for v in self.violations:
+        for v in self.get_violations():
             if hasattr(v, 'check_tuple'):
                 vs.append(v.check_tuple())
             else:
@@ -65,6 +66,16 @@ class LintedFile(namedtuple('ProtoFile', ['path', 'violations', 'time_dict', 'tr
         # Filter ignorable violations
         if filter_ignore:
             violations = [v for v in violations if not v.ignore]
+            # Ignore any rules in the ignore mask
+            if self.ignore_mask:
+                for line_no, rules in self.ignore_mask:
+                    violations = [
+                        v for v in violations
+                        if not (
+                            v.line_no() == line_no
+                            and (rules is None or v.rule_code() in rules)
+                        )
+                    ]
         return violations
 
     def num_violations(self, **kwargs):
@@ -626,6 +637,27 @@ class Linter:
         bencher("Finish parsing {0!r}".format(short_fname))
         return parsed, violations, time_dict
 
+    @staticmethod
+    def extract_ignore_from_comment(comment):
+        """Extract ignore mask entries from a comment segment."""
+        # Also trim any whitespace afterward
+        comment_content = comment.raw_trimmed().strip()
+        if comment_content.startswith('noqa'):
+            # This is an ignore identifier
+            comment_remainder = comment_content[4:]
+            if comment_remainder:
+                if not comment_remainder.startswith(':'):
+                    return SQLParseError(
+                        "Malformed 'noqa' section. Expected 'noqa: <rule>[,...]",
+                        segment=comment
+                    )
+                comment_remainder = comment_remainder[1:]
+                rules = [r.strip() for r in comment_remainder.split(',')]
+                return (comment.pos_marker.line_no, tuple(rules))
+            else:
+                return (comment.pos_marker.line_no, None)
+        return None
+
     def lint_string(self, s, fname='<string input>', verbosity=0, fix=False, config=None):
         """Lint a string.
 
@@ -643,6 +675,20 @@ class Linter:
 
         # Using the new parser, read the file object.
         parsed, vs, time_dict = self.parse_string(s=s, fname=fname, verbosity=verbosity, config=config)
+
+        # Look for comment segments which might indicate lines to ignore.
+        ignore_buff = []
+        for comment in parsed.recursive_crawl('comment'):
+            if comment.name == 'inline_comment':
+                ignore_entry = self.extract_ignore_from_comment(comment)
+                if isinstance(ignore_entry, SQLParseError):
+                    vs.append(ignore_entry)
+                elif ignore_entry:
+                    ignore_buff.append(ignore_entry)
+        if ignore_buff and verbosity >= 2:
+            verbosity_logger(
+                "Parsed noqa directives from file: {0!r}".format(ignore_buff),
+                verbosity=verbosity)
 
         templ_buff = None
         fixed_buff = None
@@ -743,7 +789,7 @@ class Linter:
 
         file_mask = (raw_buff, templ_buff, fixed_buff)
         res = LintedFile(fname, vs, time_dict, parsed,
-                         file_mask=file_mask)
+                         file_mask=file_mask, ignore_mask=ignore_buff)
 
         # This is the main command line output from linting.
         self.log(
@@ -754,27 +800,56 @@ class Linter:
         )
         return res
 
-    def paths_from_path(self, path, ignore_non_existent_files=False):
+    def paths_from_path(self, path, ignore_file_name='.sqlfluffignore', ignore_non_existent_files=False):
         """Return a set of sql file paths from a potentially more ambigious path string."""
+
+        Here we also deal with the .sqlfluffignore file if present.
+
+        """
         if not os.path.exists(path):
             if ignore_non_existent_files:
                 return []
             else:
                 raise IOError("Specified path does not exist")
 
+        # Files referred to exactly are never ignored.
         if not os.path.isdir(path):
             return [path]
 
         # If it's a directory then expand the path!
+        ignore_set = set()
         buffer = []
         for dirpath, _, filenames in os.walk(path):
             for fname in filenames:
+                fpath = os.path.join(dirpath, fname)
+                # Handle potential .sqlfluffignore files
+                if fname == ignore_file_name:
+                    with open(fpath, 'r') as fh:
+                        spec = pathspec.PathSpec.from_lines('gitwildmatch', fh)
+                    matches = spec.match_tree(dirpath)
+                    for m in matches:
+                        ignore_path = os.path.join(dirpath, m)
+                        ignore_set.add(ignore_path)
+                    # We don't need to process the ignore file any futher
+                    continue
+
+                # We won't purge files *here* because there's an edge case
+                # that the ignore file is processed after the sql file.
+
+                # Scan for remaining files
                 for ext in self.sql_exts:
                     # is it a sql file?
                     if fname.endswith(ext):
-                        # join the paths and normalise
-                        buffer.append(os.path.normpath(os.path.join(dirpath, fname)))
-        return sorted(buffer)
+                        buffer.append(fpath)
+
+        # Check the buffer for ignore items and normalise the rest.
+        filtered_buffer = []
+        for fpath in buffer:
+            if fpath not in ignore_set:
+                filtered_buffer.append(os.path.normpath(fpath))
+
+        # Return
+        return sorted(filtered_buffer)
 
     def lint_string_wrapped(self, string, fname='<string input>', verbosity=0, fix=False):
         """Lint strings directly."""