+++ src/sqlfluff/parser/segments_base.py
@@ -189,6 +189,9 @@ class BaseSegment:
     is_meta = False
     # Are we able to have non-code at the start or end?
     _can_start_end_non_code = False
+    # What should we trim off the ends to get to content
+    trim_chars = None
+    trim_start = None
 
     @property
     def name(self):
@@ -609,6 +612,18 @@ class BaseSegment:
         This raw function can be overridden, or a grammar defined
         on the underlying class.
         """
+        # Edge case, but it's possible that we have *already matched* on
+        # a previous cycle. Do should first check whether this is a case
+        # of that.
+        if len(segments) == 1 and isinstance(segments[0], cls):
+            # This has already matched. Winner.
+            parse_match_logging(cls.__name__[:10], '_match', 'SELF', parse_context=parse_context, v_level=3, symbol='+++')
+            return MatchResult.from_matched(segments)
+        elif len(segments) > 1 and isinstance(segments[0], cls):
+            parse_match_logging(cls.__name__[:10], '_match', 'SELF', parse_context=parse_context, v_level=3, symbol='+++')
+            # This has already matched, but only partially.
+            return MatchResult((segments[0],), segments[1:])
+
         if cls._match_grammar():
             # Call the private method
             m = cls._match_grammar()._match(segments=segments, parse_context=parse_context.copy(incr='match_depth'))
@@ -705,14 +720,12 @@ class BaseSegment:
     def iter_raw_seg(self):
         """Iterate raw segments, mostly for searching."""
         for s in self.segments:
-            for seg in s.iter_raw_seg():
-                yield seg
+            yield from s.iter_raw_seg()
 
     def iter_unparsables(self):
         """Iterate through any unparsables this segment may contain."""
         for s in self.segments:
-            for u in s.iter_unparsables():
-                yield u
+            yield from s.iter_unparsables()
 
     def type_set(self):
         """Return a set of the types contained, mostly for testing."""
@@ -920,6 +933,30 @@ class BaseSegment:
             pos_marker=self.pos_marker
         )
 
+    def get_child(self, seg_type):
+        """Retrieve the first of the children of this segment with matching type."""
+        for seg in self.segments:
+            if seg.type == seg_type:
+                return seg
+        return None
+
+    def get_children(self, seg_type):
+        """Retrieve the all of the children of this segment with matching type."""
+        buff = []
+        for seg in self.segments:
+            if seg.type == seg_type:
+                buff.append(seg)
+        return buff
+
+    def recursive_crawl(self, seg_type):
+        """Recursively crawl for segments of a given type."""
+        # Check this segment
+        if self.type == seg_type:
+            yield self
+        # Recurse
+        for seg in self.segments:
+            yield from seg.recursive_crawl(seg_type=seg_type)
+
 
 class RawSegment(BaseSegment):
     """This is a segment without any subsegments."""
@@ -968,6 +1005,26 @@ class RawSegment(BaseSegment):
         This is in case something tries to iterate on this segment.
         """
         return []
+
+    def raw_trimmed(self):
+        """Return a trimmed version of the raw content."""
+        raw_buff = self.raw
+        if self.trim_start:
+            for seq in self.trim_start:
+                if raw_buff.startswith(seq):
+                    raw_buff = raw_buff[len(seq):]
+        if self.trim_chars:
+            raw_buff = self.raw
+            # for each thing to trim
+            for seq in self.trim_chars:
+                # trim start
+                while raw_buff.startswith(seq):
+                    raw_buff = raw_buff[len(seq):]
+                # trim end
+                while raw_buff.endswith(seq):
+                    raw_buff = raw_buff[:-len(seq)]
+            return raw_buff
+        return raw_buff
 
     def raw_list(self):
         """Return a list of the raw content of this segment."""