+++ test/fixtures/linter/autofix/ansi/008_L022L023L024_with_clause/before.sql
@@ -1,0 +1,1 @@
+with a as(select 1), b as(select 2) select * from a join b using(z)