+++ test/fixtures/linter/select_distinct_group_by.sql
@@ -1,0 +1,1 @@
+sELECT distinct a, b FROM tbl GROUP BY a, b